altvfo
  * add code to set and read alternate vfo freq IC7300
    - command code 0x25
  * use rig data vfo initialize failed to
    correctly read start up values
  * Modify code to copy freq, mode and bandwidth A->B
  * Change all Icom class vfo set/get methods
  * Change A=B behavior; active vfo --> inactive vfo
    independent of whether A or B is active.  This
    behavior emulates the Icom series of transceivers.

+++ src/include/tod_clock.h
@@ -22,7 +22,7 @@
 #ifndef TOD_H_
 #define TOD_H_
 
-extern unsigned long zmsec();
+extern int   zmsec();
 extern char* zdate();
 extern char* ztime();
 extern char* zshowtime();
